chore(genja-core-derive): Moved the genja-core dependant tests from genja-core-derive into genja-core.
