+++ genja-core-derive/Cargo.toml
@@ -14,9 +14,6 @@ quote = "1.0.40"
 syn = "2.0.106"
 
 [dev-dependencies]
-async-trait = "0.1.89"
-genja-core = { path = "../genja-core" }
-serde_json = "1.0.142"
 trybuild = "1.0.113"
 
 [lib]
