+++ genja-core-derive/tests/compile.rs
@@ -1,11 +1,11 @@
 #[test]
-fn task_derive_passes_supported_inputs() {
+fn deref_derive_passes_supported_inputs() {
     let test_cases = trybuild::TestCases::new();
-    test_cases.pass("tests/ui/pass/*.rs");
+    test_cases.pass("tests/ui/pass/deref_*.rs");
 }
 
 #[test]
-fn task_derive_rejects_unsupported_inputs() {
+fn deref_derive_rejects_unsupported_inputs() {
     let test_cases = trybuild::TestCases::new();
-    test_cases.compile_fail("tests/ui/fail/*.rs");
+    test_cases.compile_fail("tests/ui/fail/deref_*.rs");
 }
