+++ genja-core/Cargo.toml
@@ -30,3 +30,4 @@ regex = "1.12.2"
 tempfile = "3.24.0"
 tracing-rolling-file = { version = "0.1.3", features = ["non-blocking"] }
 tracing-subscriber = "0.3"
+trybuild = "1.0.113"
