+++ genja-core/tests/ui/fail/task_unknown_struct_attribute.stderr
