+++ genja-core/tests/ui/pass/task_minimal_string.rs
