+++ genja-core/tests/ui/pass/task_static_name_and_processors.rs
